fix(skole): safe int cast and split verifyTeacherOwnsClassroom log conditions
Use Math.toIntExact instead of raw cast in ClassroomService and SchoolService.
Split the two conditions in verifyTeacherOwnsClassroom so logs correctly
distinguish "classroom not found" from "teacher not authorized".

+++ backend/src/main/java/no/ntnu/idatt2106/nettdetektivene/service/SchoolService.java
@@ -95,7 +95,7 @@ public class SchoolService {
             return List.of();
         }
         Long schoolId = teacher.getSchool().getId();
-        int totalTasks = (int) taskRepository.count();
+        int totalTasks = Math.toIntExact(taskRepository.count());
         return classroomRepository.findBySchool_Id(schoolId).stream()
             .map(classroom -> toSchoolClassroomSummary(classroom, totalTasks))
             .toList();
